+++ src/main/java/racingcar/dao/game/ConsoleGameDao.java
@@ -1,4 +1,4 @@
-package racingcar.dao;
+package racingcar.dao.game;
 
 import java.util.List;
 import racingcar.entity.GameEntity;
