+++ src/main/java/racingcar/dao/game/WebGameDao.java
@@ -1,4 +1,4 @@
-package racingcar.dao;
+package racingcar.dao.game;
 
 import java.sql.PreparedStatement;
 import java.util.List;
