+++ src/main/java/racingcar/service/RacingCarService.java
@@ -6,9 +6,9 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import racingcar.dao.GameDao;
-import racingcar.dao.ParticipatesDao;
-import racingcar.dao.PlayerDao;
+import racingcar.dao.game.GameDao;
+import racingcar.dao.participates.ParticipatesDao;
+import racingcar.dao.player.PlayerDao;
 import racingcar.domain.Cars;
 import racingcar.domain.GameManager;
 import racingcar.domain.GameRound;
